Update ae.py
I was reading your code as a reference. Then I found that you are converting array to list for converting it to floatX dtype, but is is much efficient not to convert to list then to array rather we can directly convert to floatX. 
I have also added corruption quantity.